move summarize tests to their own file
